+++ instrumentation/opentelemetry-instrumentation-redis/src/opentelemetry/instrumentation/redis/__init__.py
@@ -56,7 +56,6 @@ from opentelemetry.instrumentation.utils import unwrap
 
 _DEFAULT_SERVICE = "redis"
 _RAWCMD = "db.statement"
-_CMD = "redis.command"
 
 
 def _set_connection_attributes(span, conn):
@@ -71,8 +70,13 @@ def _set_connection_attributes(span, conn):
 def _traced_execute_command(func, instance, args, kwargs):
     tracer = getattr(redis, "_opentelemetry_tracer")
     query = _format_command_args(args)
+    name = ""
+    if len(args) > 0 and args[0]:
+        name = args[0]
+    else:
+        name = instance.connection_pool.connection_kwargs.get("db", 0)
     with tracer.start_as_current_span(
-        _CMD, kind=trace.SpanKind.CLIENT
+        name, kind=trace.SpanKind.CLIENT
     ) as span:
         if span.is_recording():
             span.set_attribute("service", tracer.instrumentation_info.name)
@@ -88,8 +92,10 @@ def _traced_execute_pipeline(func, instance, args, kwargs):
     cmds = [_format_command_args(c) for c, _ in instance.command_stack]
     resource = "\n".join(cmds)
 
+    span_name = " ".join([args[0] for args, _ in instance.command_stack])
+
     with tracer.start_as_current_span(
-        _CMD, kind=trace.SpanKind.CLIENT
+        span_name, kind=trace.SpanKind.CLIENT
     ) as span:
         if span.is_recording():
             span.set_attribute("service", tracer.instrumentation_info.name)