+++ instrumentation/opentelemetry-instrumentation-redis/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Update redis instrumentation to follow semantic conventions
+  ([#184](https://github.com/open-telemetry/opentelemetry-python-contrib/pull/184))
+
 ## Version 0.13b0
 
 Released 2020-09-17